+++ src/Document.php
@@ -54,10 +54,14 @@ class Document implements DocumentInterface
 
     /**
      * @param \Psr\Http\Message\ResponseInterface|null $response
+     *
+     * @return $this
      */
-    public function setResponse(ResponseInterface $response = null)
+    public function setResponse(? ResponseInterface $response)
     {
         $this->response = $response;
+
+        return $this;
     }
 
     /**
@@ -70,10 +74,14 @@ class Document implements DocumentInterface
 
     /**
      * @param \Swis\JsonApi\Client\ErrorCollection $errors
+     *
+     * @return $this
      */
     public function setErrors(ErrorCollection $errors)
     {
         $this->errors = $errors;
+
+        return $this;
     }
 
     /**
@@ -103,7 +111,7 @@ class Document implements DocumentInterface
     /**
      * @param \Swis\JsonApi\Client\Collection $included
      *
-     * @return static
+     * @return $this
      */
     public function setIncluded(Collection $included)
     {
@@ -122,10 +130,14 @@ class Document implements DocumentInterface
 
     /**
      * @param \Swis\JsonApi\Client\Jsonapi|null $jsonapi
+     *
+     * @return $this
      */
-    public function setJsonapi(Jsonapi $jsonapi = null)
+    public function setJsonapi(? Jsonapi $jsonapi)
     {
         $this->jsonapi = $jsonapi;
+
+        return $this;
     }
 
     /**
@@ -139,7 +151,7 @@ class Document implements DocumentInterface
     /**
      * @param \Swis\JsonApi\Client\Interfaces\DataInterface $data
      *
-     * @return static
+     * @return $this
      */
     public function setData(DataInterface $data)
     {