+++ src/Interfaces/DocumentInterface.php
@@ -17,14 +17,35 @@ interface DocumentInterface extends \JsonSerializable
     public function getResponse(): ? ResponseInterface;
 
     /**
+     * @param \Psr\Http\Message\ResponseInterface|null $response
+     *
+     * @return $this
+     */
+    public function setResponse(? ResponseInterface $response);
+
+    /**
      * @return \Swis\JsonApi\Client\Interfaces\DataInterface
      */
     public function getData();
 
     /**
+     * @param \Swis\JsonApi\Client\Interfaces\DataInterface $data
+     *
+     * @return $this
+     */
+    public function setData(DataInterface $data);
+
+    /**
      * @return \Swis\JsonApi\Client\ErrorCollection
      */
     public function getErrors(): ErrorCollection;
+
+    /**
+     * @param \Swis\JsonApi\Client\ErrorCollection $errors
+     *
+     * @return $this
+     */
+    public function setErrors(ErrorCollection $errors);
 
     /**
      * @return bool
@@ -37,9 +58,23 @@ interface DocumentInterface extends \JsonSerializable
     public function getMeta(): ? Meta;
 
     /**
+     * @param \Swis\JsonApi\Client\Meta|null $meta
+     *
+     * @return $this
+     */
+    public function setMeta(? Meta $meta);
+
+    /**
      * @return \Swis\JsonApi\Client\Links|null
      */
     public function getLinks(): ? Links;
+
+    /**
+     * @param \Swis\JsonApi\Client\Links|null $links
+     *
+     * @return $this
+     */
+    public function setLinks(? Links $links);
 
     /**
      * @return mixed
@@ -47,9 +82,23 @@ interface DocumentInterface extends \JsonSerializable
     public function getIncluded(): Collection;
 
     /**
+     * @param \Swis\JsonApi\Client\Collection $included
+     *
+     * @return $this
+     */
+    public function setIncluded(Collection $included);
+
+    /**
      * @return \Swis\JsonApi\Client\Jsonapi|null
      */
     public function getJsonapi(): ? Jsonapi;
+
+    /**
+     * @param \Swis\JsonApi\Client\Jsonapi|null $jsonapi
+     *
+     * @return $this
+     */
+    public function setJsonapi(? Jsonapi $jsonapi);
 
     /**
      * @return array