+++ src/Parsers/LinksParser.php
@@ -1,20 +1,36 @@
 <?php
 
-namespace Swis\JsonApi\Client\JsonApi;
+namespace Swis\JsonApi\Client\Parsers;
 
 use Art4\JsonApiClient\Meta as JsonApiMeta;
 use Swis\JsonApi\Client\Link;
 use Swis\JsonApi\Client\Links;
 use Swis\JsonApi\Client\Meta;
 
+/**
+ * @internal
+ */
 class LinksParser
 {
+    /**
+     * @var \Swis\JsonApi\Client\Parsers\MetaParser
+     */
+    private $metaParser;
+
+    /**
+     * @param \Swis\JsonApi\Client\Parsers\MetaParser $metaParser
+     */
+    public function __construct(MetaParser $metaParser)
+    {
+        $this->metaParser = $metaParser;
+    }
+
     /**
      * @param array $links
      *
      * @return \Swis\JsonApi\Client\Links
      */
-    public function parse(array $links)
+    public function parse(array $links): Links
     {
         return new Links(
             array_map(
@@ -27,7 +43,7 @@ class LinksParser
     }
 
     /**
-     * @param \Art4\JsonApiClient\Link|string $link
+     * @param \Art4\JsonApiClient\DocumentLink|\Art4\JsonApiClient\ErrorLink|\Art4\JsonApiClient\Link|\Art4\JsonApiClient\RelationshipLink|\Art4\JsonApiClient\ResourceItemLink|string $link
      *
      * @return \Swis\JsonApi\Client\Link
      */
@@ -47,6 +63,6 @@ class LinksParser
      */
     private function buildMeta(JsonApiMeta $meta): Meta
     {
-        return new Meta($meta->asArray(true));
+        return $this->metaParser->parse($meta);
     }
 }